Move api way version test to integration tests
